Adjust ratelimit and quota in manifest_template.yml (#104)
* Adjust ratelimit and quota in manifest_template.yml

This commit updates the values of 'ratelimit' and 'quota' for different environments in the manifest_template.yml file. The changes are aimed at correctly configuring the request limits for each environment, thus improving the system's performance and reliability.

* Enable app quota and spikeArrest in manifest_template.yml

The previous comments in the manifest_template.yml file were uncommented to enable app quota and spikeArrest. This will ensure the correct quota limits and spike arrest configuration are set, which is essential for maintaining the efficiency and stability of different environments.

---------

Co-authored-by: igba.ujege <igba.ujege@bjss.com>