Add parameters to ApplyRateLimiting flow callout (#106)
* Add parameters to ApplyRateLimiting flow callout

A 'Parameters' section was added to the 'FlowCallout.ApplyRateLimiting' in apiproxy/policies. This includes a parameter named 'verifyApiKeyPolicyName' with the value 'VerifyApiKey.FromHeader'. This will ensure that the API key is verified from the header, improving the security measure of the process.

* Changed apikey case

* Update display name in FlowCallout.ApplyRateLimiting

The display name of the FlowCallout "ApplyRateLimiting" in the "live" proxy's policy was updated. The change made was specifically to improve readability without affecting functionality.

---------

Co-authored-by: igba.ujege <igba.ujege@bjss.com>